Add migration and seeder files

+++ database/migrations/2025_07_09_214231_create_mata_kuliah_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('mata_kuliah', function (Blueprint $table) {
             $table->bigIncrements('id_mk');
-            $table->integer('id_ps');
+            $table->unsignedBigInteger('id_ps');
             $table->string('nama_matkul');
             $table->integer('jumlah_sks');
             $table->integer('semester');
