+++ database/migrations/2025_07_09_214303_create_profil_lulusan_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('profil_lulusan', function (Blueprint $table) {
             $table->bigIncrements('id_pl');
-            $table->integer('id_ps');
+            $table->unsignedBigInteger('id_ps');
             $table->string('profil_lulusan');
             $table->text('desc');
