+++ database/migrations/2025_07_10_004423_create_user_role_map_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('user_role_map', function (Blueprint $table) {
             $table->bigIncrements('id_user_role');
-            $table->integer('id_role');
-            $table->integer('id_user');
+            $table->unsignedBigInteger('id_role');
+            $table->unsignedBigInteger('id_user');
 
             $table->foreign('id_role')->references('id_role')->on('role');
             $table->foreign('id_user')->references('id_user')->on('user');
