+++ database/migrations/2025_07_10_005430_create_sub_cpmk_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('sub_cpmk', function (Blueprint $table) {
             $table->bigIncrements('id_sub_cpmk');
-            $table->integer('id_cpmk');
+            $table->unsignedBigInteger('id_cpmk');
             $table->string('nama_kode_sub_cpmk');
             $table->integer('kode_sub_cpmk');
             $table->text('desc');
