+++ database/migrations/2025_07_10_005440_create_rps_detail_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('rps_detail', function (Blueprint $table) {
             $table->bigIncrements('id_rps_detail');
-            $table->integer('id_rps');
-            $table->integer('id_sub_cpmk');
+            $table->unsignedBigInteger('id_rps');
+            $table->unsignedBigInteger('id_sub_cpmk');
             $table->integer('minggu');
             $table->decimal('penilaian');
             $table->decimal('bobot');
