+++ database/migrations/2025_07_10_010324_create_user_mata_kuliah_map_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('user_mata_kuliah_map', function (Blueprint $table) {
             $table->bigIncrements('id_user_mk');
-            $table->integer('id_user');
-            $table->integer('id_mk');
+            $table->unsignedBigInteger('id_user');
+            $table->unsignedBigInteger('id_mk');
 
             $table->foreign('id_user')->references('id_user')->on('user');
             $table->foreign('id_mk')->references('id_mk')->on('mata_kuliah');
