+++ database/migrations/2025_07_10_010436_create_mahasiswa_cpl_map_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('mahasiswa_cpl_map', function (Blueprint $table) {
             $table->bigIncrements('id_mhs_cpl');
-            $table->integer('id_mhs');
-            $table->integer('id_cpl');
+            $table->unsignedBigInteger('id_mhs');
+            $table->unsignedBigInteger('id_cpl');
 
             $table->foreign('id_mhs')->references('id_mhs')->on('mahasiswa');
             $table->foreign('id_cpl')->references('id_cpl')->on('cpl');
